+++ src/common/Formatter.h
@@ -85,128 +85,5 @@ namespace ceph {
     }
   };
 
-  class JSONFormatter : public Formatter {
-  public:
-    JSONFormatter(bool p = false);
-
-    void flush(std::ostream& os);
-    void reset();
-    virtual void open_array_section(const char *name);
-    void open_array_section_in_ns(const char *name, const char *ns);
-    void open_object_section(const char *name);
-    void open_object_section_in_ns(const char *name, const char *ns);
-    void close_section();
-    void dump_unsigned(const char *name, uint64_t u);
-    void dump_int(const char *name, int64_t u);
-    void dump_float(const char *name, double d);
-    void dump_string(const char *name, const std::string& s);
-    std::ostream& dump_stream(const char *name);
-    void dump_format_va(const char *name, const char *ns, bool quoted, const char *fmt, va_list ap);
-    int get_len() const;
-    void write_raw_data(const char *data);
-
-  private:
-
-    struct json_formatter_stack_entry_d {
-      int size;
-      bool is_array;
-      json_formatter_stack_entry_d() : size(0), is_array(false) { }
-    };
-
-    bool m_pretty;
-    void open_section(const char *name, bool is_array);
-    void print_quoted_string(const std::string& s);
-    void print_name(const char *name);
-    void print_comma(json_formatter_stack_entry_d& entry);
-    void finish_pending_string();
-
-    std::stringstream m_ss, m_pending_string;
-    std::list<json_formatter_stack_entry_d> m_stack;
-    bool m_is_pending_string;
-  };
-
-  class XMLFormatter : public Formatter {
-  public:
-    static const char *XML_1_DTD;
-    XMLFormatter(bool pretty = false);
-
-    void flush(std::ostream& os);
-    void reset();
-    void open_array_section(const char *name);
-    void open_array_section_in_ns(const char *name, const char *ns);
-    void open_object_section(const char *name);
-    void open_object_section_in_ns(const char *name, const char *ns);
-    void close_section();
-    void dump_unsigned(const char *name, uint64_t u);
-    void dump_int(const char *name, int64_t u);
-    void dump_float(const char *name, double d);
-    void dump_string(const char *name, const std::string& s);
-    std::ostream& dump_stream(const char *name);
-    void dump_format_va(const char *name, const char *ns, bool quoted, const char *fmt, va_list ap);
-    int get_len() const;
-    void write_raw_data(const char *data);
-
-    /* with attrs */
-    void open_array_section_with_attrs(const char *name, const FormatterAttrs& attrs);
-    void open_object_section_with_attrs(const char *name, const FormatterAttrs& attrs);
-    void dump_string_with_attrs(const char *name, const std::string& s, const FormatterAttrs& attrs);
-  private:
-    void open_section_in_ns(const char *name, const char *ns, const FormatterAttrs *attrs);
-    void finish_pending_string();
-    void print_spaces();
-    static std::string escape_xml_str(const char *str);
-    void get_attrs_str(const FormatterAttrs *attrs, std::string& attrs_str);
-
-    std::stringstream m_ss, m_pending_string;
-    std::deque<std::string> m_sections;
-    bool m_pretty;
-    std::string m_pending_string_name;
-  };
-
-  class TableFormatter : public Formatter {
-  public:
-    TableFormatter(bool keyval = false);
-
-    void flush(std::ostream& os);
-    void reset();
-    virtual void open_array_section(const char *name);
-    void open_array_section_in_ns(const char *name, const char *ns);
-    void open_object_section(const char *name);
-    void open_object_section_in_ns(const char *name, const char *ns);
-
-    void open_array_section_with_attrs(const char *name, const FormatterAttrs& attrs);
-    void open_object_section_with_attrs(const char *name, const FormatterAttrs& attrs);
-
-    void close_section();
-    void dump_unsigned(const char *name, uint64_t u);
-    void dump_int(const char *name, int64_t u);
-    void dump_float(const char *name, double d);
-    void dump_string(const char *name, const std::string& s);
-    void dump_format_va(const char *name, const char *ns, bool quoted, const char *fmt, va_list ap);
-    void dump_string_with_attrs(const char *name, const std::string& s, const FormatterAttrs& attrs);
-    std::ostream& dump_stream(const char *name);
-
-    int get_len() const;
-    void write_raw_data(const char *data);
-    void get_attrs_str(const FormatterAttrs *attrs, std::string& attrs_str);
-
-  private:
-    void open_section_in_ns(const char *name, const char *ns, const FormatterAttrs *attrs);
-    std::vector< std::vector<std::pair<std::string, std::string> > > m_vec;
-    std::stringstream m_ss;
-    size_t m_vec_index(const char* name);
-    std::string get_section_name(const char* name);
-    void finish_pending_string();
-    std::string m_pending_name;
-    bool m_keyval;
-
-    int m_section_open;
-    std::vector< std::string > m_section;
-    std::map<std::string, int> m_section_cnt;
-    std::vector<size_t> m_column_size;
-    std::vector< std::string > m_column_name;
-  };
-
-
 }
 #endif
