+++ src/common/Makefile.am
@@ -45,6 +45,9 @@ libcommon_internal_la_SOURCES = \
 	common/simple_spin.cc \
 	common/Thread.cc \
 	common/Formatter.cc \
+	common/XMLFormatter.cc \
+	common/JSONFormatter.cc \
+	common/TableFormatter.cc \
 	common/HeartbeatMap.cc \
 	common/config.cc \
 	common/utf8.c \
@@ -178,6 +181,9 @@ noinst_HEADERS += \
 	common/DecayCounter.h \
 	common/Finisher.h \
 	common/Formatter.h \
+	common/JSONFormatter.h \
+	common/TableFormatter.h \
+	common/XMLFormatter.h \
 	common/perf_counters.h \
 	common/OutputDataSocket.h \
 	common/admin_socket.h \
