+++ src/mon/ConfigKeyService.cc
@@ -24,6 +24,7 @@
 #include "common/config.h"
 #include "common/cmdparse.h"
 #include "common/errno.h"
+#include "common/JSONFormatter.h"
 
 #define dout_subsys ceph_subsys_mon
 #undef dout_prefix
