+++ src/os/KeyValueStore.cc
@@ -48,6 +48,8 @@
 #include "common/safe_io.h"
 #include "common/perf_counters.h"
 #include "common/sync_filesystem.h"
+#include "common/Formatter.h"
+#include "common/JSONFormatter.h"
 
 #ifdef HAVE_KINETIC
 #include "KineticStore.h"
