+++ src/os/MemStore.cc
@@ -26,6 +26,8 @@
 #include "include/unordered_map.h"
 #include "include/memory.h"
 #include "common/errno.h"
+#include "common/Formatter.h"
+#include "common/JSONFormatter.h"
 #include "MemStore.h"
 
 #define dout_subsys ceph_subsys_filestore
