+++ src/rgw/rgw_admin.cc
@@ -15,6 +15,8 @@ using namespace std;
 #include "common/config.h"
 #include "common/ceph_argparse.h"
 #include "common/Formatter.h"
+#include "common/XMLFormatter.h"
+#include "common/JSONFormatter.h"
 #include "common/errno.h"
 
 #include "global/global_init.h"
