+++ src/rgw/rgw_cors_s3.h
@@ -22,6 +22,7 @@
 
 #include <include/types.h>
 #include <common/Formatter.h>
+#include <common/XMLFormatter.h>
 #include "rgw_xml.h"
 #include "rgw_cors.h"
