+++ src/test/bench/small_io_bench_dumb.cc
@@ -16,6 +16,7 @@
 #include <fstream>
 
 #include "common/Formatter.h"
+#include "common/JSONFormatter.h"
 
 #include "bencher.h"
 #include "rados_backend.h"
