+++ src/test/bench/small_io_bench_rbd.cc
@@ -13,6 +13,7 @@
 #include <fstream>
 
 #include "common/Formatter.h"
+#include "common/JSONFormatter.h"
 
 #include "bencher.h"
 #include "rbd_backend.h"
