+++ src/test/common/test_tableformatter.cc
@@ -1,6 +1,6 @@
 #include "gtest/gtest.h"
 
-#include "common/Formatter.h"
+#include "common/TableFormatter.h"
 #include <iostream>
 #include <sstream>
 #include <string>
