+++ src/test/crush/crush.cc
@@ -13,6 +13,7 @@
 
 #include "include/stringify.h"
 #include "common/ceph_argparse.h"
+#include "common/JSONFormatter.h"
 #include "global/global_init.h"
 #include "global/global_context.h"
