+++ src/test/encoding/ceph_dencoder.cc
@@ -5,6 +5,7 @@
 #include "include/ceph_features.h"
 #include "common/ceph_argparse.h"
 #include "common/Formatter.h"
+#include "common/JSONFormatter.h"
 #include "common/errno.h"
 #include "msg/Message.h"
 #include "include/assert.h"
