+++ src/tools/ceph_monstore_tool.cc
@@ -16,7 +16,7 @@
 #include <stdlib.h>
 #include <string>
 
-#include "common/Formatter.h"
+#include "common/JSONFormatter.h"
 #include "common/errno.h"
 
 #include "global/global_init.h"
