+++ src/tools/cephfs/EventOutput.cc
@@ -16,6 +16,7 @@
 #include <fstream>
 
 #include "common/errno.h"
+#include "common/JSONFormatter.h"
 #include "mds/mdstypes.h"
 #include "mds/events/EUpdate.h"
 #include "mds/LogEvent.h"
